+++ lib/graphlient/client.rb
@@ -5,9 +5,10 @@ module Graphlient
   class Client
     attr_accessor :uri, :options
 
-    def initialize(url, options = {})
+    def initialize(url, options = {}, &_block)
       @url = url
       @options = options.dup
+      yield self if block_given?
     end
 
     def query(&block)
@@ -21,11 +22,14 @@ module Graphlient
       raise Graphlient::Errors::Client.new(e.message, e)
     end
 
+    def http(&block)
+      @http ||= Adapters::FaradayAdapter.new(@url, headers: @options[:headers], &block)
+    end
+
     private
 
     def client
       @client ||= begin
-        http = Adapters::FaradayAdapter.new(@url, headers: @options[:headers])
         # Fetch latest schema on init, this will make a network request
         schema = GraphQL::Client.load_schema(http)
         # However, it's smart to dump this to a JSON file and load from disk
@@ -34,7 +38,7 @@ module Graphlient
         #   GraphQL::Client.dump_schema(SWAPI::HTTP, "path/to/schema.json")
         #
         # Schema = GraphQL::Client.load_schema("path/to/schema.json")
-        @client = GraphQL::Client.new(schema: schema, execute: http)
+        GraphQL::Client.new(schema: schema, execute: http)
       end
     end
   end