+++ lib/graphlient/errors.rb
@@ -1,1 +1,3 @@
+require_relative 'errors/error'
 require_relative 'errors/client'
+require_relative 'errors/server'