+++ spec/support/types/invoice_type.rb
@@ -1,0 +1,6 @@
+InvoiceType = GraphQL::ObjectType.define do
+  name 'Invoice'
+  description 'An Invoice'
+  field :id, !types.ID
+  field :fee_in_cents, types.Int
+end